bot/requirements.txt
aiogram>=3.4
aiohttp>=3.9
sqlalchemy>=2.0
aiosqlite>=0.19
asyncpg>=0.29
alembic>=1.13
redis>=5.0
apscheduler>=3.10
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.1
